Factor out call related visitor code
